chi.gof1 {Rstat}	R Documentation

Goodness of Fit Test from a Table

Description

Goodness of fit test from a table, when the probability is given.

Usage

chi.gof1(x, p, alp = 0.05, dig = 2, dig2 = 4)

Arguments

x
Table of frequency

p
Probability vector

alp
Level of significance, Default: 0.05

dig
Number of digits below the decimal point for the table, Default: 2

dig2
Number of digits below the decimal point for the results, Default: 4

Value

list(stat=test stat, df=degree of freedom, crv=critical value, pv=p-value, tab=table)

Examples

# Goodness-of-fit Test
x = c(31,26,22,18,13,10)
ans = chi.gof1(x)

[Package Rstat version 0.0.0.1 Index]