civar2.sim {Rstat}	R Documentation

Simulate the Confidence Interval for a the ratio of Variances

Description

Simulate the Confidence Interval for a the ratio of Population Variances

Usage

civar2.sim(n1, n2, sig1, sig2, alp = 0.05, N = 100, seed = 9857,
  dig = 4, plot = TRUE)

Arguments

n1
Sample size of population1

n2
Sample size of population2

sig1
Standard deviation of population1

sig2
Standard deviation of population2

alp
Level of significance, Default: 0.05

N
Number of iterations, Default: 100

seed
Seed value for generating random numbers, Default: 9857

dig
Number of digits below the decimal point, Default: 4

plot
Logical value for plot, Default: TRUE

Value

None.

Examples

civar2.sim(n1=25, n2=16, sig1=sqrt(8), sig2=2)
civar2.sim(n1=25, n2=16, sig1=sqrt(8), sig2=2, N=10000, plot=F)

[Package Rstat version 0.0.0.1 Index]