cont.jcdf {Rstat}	R Documentation

Joint CDF of Two Continuous Random Variables

Description

Joint Cumulative Distribution Function of Two Continuous Random Variables

Usage

cont.jcdf(FUN, xs, ys, lo1 = -Inf, lo2 = -Inf)

Arguments

FUN
Continuous joint PDF function

xs
Specific values of X for displaying the cumulative probability

ys
Specific values of Y for displaying the cumulative probability

lo1
Lower limit of X, Default: -Inf

lo2
Lower limit of Y, Default: -Inf

Value

Cumulative probability

Examples

pdf = function(x, y) (x+y)*(x>0 & x<1)*(y>0 & y<1)
cont.jcdf(pdf, Inf, Inf)
cont.jcdf(pdf, 0.5, 0.5)

[Package Rstat version 0.0.0.1 Index]