cont.marg2 {Rstat}	R Documentation

Marginal PDF Plot of Two Continuous Random Variables

Description

Marginal Probability Distribution Plot of Two Continuous Random Variables

Usage

cont.marg2(FUN, lo1, up1, lo2, up2, xs, ys)

Arguments

FUN
Continuous joint PDF function

lo1
Lower limit of x-axis

up1
Upper limit of x-axis

lo2
Lower limit of y-axis

up2
Upper limit of y-axis

xs
Specific value of X for displaying the probability density

ys
Specific value of Y for displaying the probability density

Value

Marginal PDF

Examples

pdf = function(x, y) 2/7*(2*x+5*y)*(x>=0 & x<=1)*(y>=0 & y<=1)
cont.marg2(pdf, -0.2, 1.2, -0.2, 1.2, 0:2/2, 0:2/2)

[Package Rstat version 0.0.0.1 Index]