cross.test {Rstat}	R Documentation

Test of a Cross Table

Description

Test of Homogeniety/Independence from a Cross Table

Usage

cross.test(x, v1, v2, alp = 0.05, dig = 2, dig2 = 4, prt = TRUE)

Arguments

x
Cross Table

v1
Row variable name (optional)

v2
Column variable name (optional)

alp
Level of significance, Default: 0.05

dig
Number of digits below the decimal point for the table, Default: 2

dig2
Number of digits below the decimal point for the results, Default: 4

prt
Logical, whether to print tables in detail, Default: TRUE

Value

list(stat=test stat, df=df, crv=critical value, pv=p-value, exp=exp table, ress=residual square)

Examples

data(exa12_7)
x=exa12_7
rownames(x)=c("Kor", "Eng", "Math", "Etc")
colnames(x)=c("Sam", "Pub", "Exp", "Sal", "Etc")
cross.test(x)

[Package Rstat version 0.0.0.1 Index]