disc.joint2 {Rstat}	R Documentation

Joint PDF of Two Discrete Random variable

Description

Joint Probability Distribution of Two Discrete Random variable

Usage

disc.joint2(X, Y, prt = TRUE, plot = FALSE, dig = 4, dig2 = 3, ep = 0)

Arguments

X
Sample space vector of the first random variable

Y
Sample space vector of the second random variable

prt
Print the joint frequency and probability? Default: TRUE

plot
Plot the joint PDF? Default: FALSE

dig
Number of digits below the decimal point in the console, Default: 4

dig2
Number of digits below the decimal point in the graph, Default: 3

ep
Minimum value for displaying the joint probability, Default: 0

Value

Joint PDF

Examples

S = rolldie2(4)
X = apply(S, 1, max)
Y = apply(S, 1, min)
disc.joint2(X, Y)
library(scatterplot3d)
disc.joint2(X, Y, prt=FALSE, plot=TRUE)

[Package Rstat version 0.0.0.1 Index]