meanchar.plot {Rstat}	R Documentation

Operating Characteristic Curves of the Test for a Mean

Description

Operating Characteristic Curves of the Test for a Population Mean

Usage

meanchar.plot(mu0, mu1, sig, nv, alp = 0.05, prng, side = "two", mt,
  dig = 4)

Arguments

mu0
Population mean value under the null hypothesis

mu1
Vector of population mean values for which the power should be calculated

sig
Population standard deviation

nv
Sample size vector

alp
Level of significance, Default: 0.05

prng
Range of x-axis

side
Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'

mt
Graph title

dig
Number of digits below the decimal point, Default: 4

Value

None.

Examples

n = c(10, 30, 50, 100)
meanchar.plot(mu0=100, mu1=98, sig=5, nv=n, side="low")
meanchar.plot(mu0=100, mu1=98, sig=5, nv=n, side="two")

[Package Rstat version 0.0.0.1 Index]