normtest.plot {Rstat}	R Documentation

Plot the PDF of Test Statistic under the Normal Distribution

Description

Plot the PDF of Test Statistic under the Normal Distribution

Usage

normtest.plot(md, mu0 = 0, se = 1, prng = c(-4, 4), side = "two",
  xlab = "Sample Mean", pvout = TRUE)

Arguments

md
Difference of sample means or test statistic

mu0
Difference of population means under the null hypothesis, Default: 0

se
Standard error of the difference of sample means, Default: 1

prng
Range of x-axis, Default: c(-4, 4)

side
Type of the alternative hypothesis, Default: 'two'

xlab
Label of x-axis, Default: 'Sample Mean'

pvout
Print p-value? Default: TRUE

Value

None.

Examples

normtest.plot(11.5-10.4, se=sqrt(2.2/20+2.4/25))
normtest.plot(1.93, xlab="Test Statistic")

[Package Rstat version 0.0.0.1 Index]